[refactor] Extract StageControlWidget from the Movement tab (FIB-783)
The second half of the split, and the last of it. FibsemMovementWidget goes from
602 lines to 168: it is now a container that lays out the four panels and forwards
the calls other windows make on the tab.

Grouped by job before moving anything, so the cut had a reason:

    137  canvas double-click (guards, dispatch, worker, teardown)
     72  move actions + workers
     47  progress / status
     22  acquisition policy
    158  host facade + layout
     96  setup_connections, spanning all of them

I considered pulling the canvas double-click out separately -- it is the largest
coherent unit and carries the FIB-329 teardown. It stays: `_click_to_move_available`
gates on `pushButton_move.isEnabled()`, so the enabled state is one thing owned
jointly by the Move button, the orientation buttons and click-to-move. Splitting it
would mean sharing that state across two widgets.

StageControlWidget takes the host explicitly rather than reaching through
`self.parent`, because it is no longer the widget the host parents. It keeps every
coupling: the image widget (the interaction handshake, the live image a click is
resolved against, the acquisition-finished signal), the optional milling widget,
and the quad-view controller. The container keeps the parent assert, since it is
still what the hosts construct.

Six test files re-pointed. Four only needed their fixture to yield
`widget.control_widget` -- the bodies are untouched, which is the equivalence
argument. `test_viewer_less_widgets` gains a `_control_widget(host)` seam beside
`_movement_widget`, and its wiring guard now spans both halves; its docstring says
so, because "a method truncated anywhere" now means either setup_connections.

Verified:

- 131 tests green across the nine movement-related files.
- The stage panel renders pixel-identically before and after: 0 differing pixels
  at 429x313. The position half shifted a column by 5px; this one does not.
- The real AutoLamellaUI built offscreen against a Demo microscope and driven
  through update_ui, the refresh button, move_to_position and teardown. No Qt
  warnings.
- The desktop application launched, connected and stayed up with no errors.
