+++ tests/ui/test_movement_progress.py
@@ -63,7 +63,7 @@ def movement(qapp):
     host = _CanvasHost()
     _image_widget(host)
     widget = _movement_widget(host)
-    yield widget
+    yield widget.control_widget
     host.deleteLater()
